Start abstracting helpers to reduce duplication
