+++ support-frontend/assets/pages/digital-subscription-checkout/components/endSummary/endSummarySelector.js
@@ -3,22 +3,12 @@
 import { type ProductPrice, getProductPrice } from 'helpers/productPrice/productPrices';
 import type { CheckoutState } from 'helpers/subscriptionsForms/subscriptionCheckoutReducer';
 import { getBillingDescription } from 'helpers/productPrice/priceDescriptionsDigital';
+import { giftPeriod } from '../helpers';
 
 const getPromotion = (productPrice: ProductPrice): string | null =>
   (productPrice.promotions && productPrice.promotions.length > 0
     ? productPrice.promotions[0].description
     : null);
-
-const giftPeriod = {
-  Annual: {
-    period: '12 month gift subscription',
-    cost: 'You\'ll pay £119 for 12 months',
-  },
-  Monthly: {
-    period: '3 month gift subscription',
-    cost: 'You\'ll pay £35.95 for 3 months',
-  },
-};
 
 function mapStateToProps(state: CheckoutState) {
   const {
