+++ support-frontend/assets/pages/digital-subscription-checkout/components/helpers.js
@@ -1,0 +1,15 @@
+// @flow
+
+const giftPeriod = {
+  Annual: {
+    period: '12 month gift subscription',
+    cost: 'You\'ll pay £119 for 12 months',
+  },
+  Monthly: {
+    period: '3 month gift subscription',
+    cost: 'You\'ll pay £35.95 for 3 months',
+  },
+};
+
+
+export { giftPeriod };
